% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROvis-package.R
\docType{package}
\name{ROvis-package}
\alias{ROvis}
\alias{ROvis-package}
\title{ROvis: Create data visualisations in Rijksoverheid style using a set of R packages}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

ROvis is an R package that provides tools to design graphs and reports in the Rijksoverheid style.
}
\author{
\strong{Maintainer}: ROvis team \email{spin@rivm.nl}

Authors:
\itemize{
  \item ROvis team \email{spin@rivm.nl}
}

}
\keyword{internal}
